fix(🎨): consolidate color processing in the scenegraph
